NLP/requirements.txt
click
pandas
wordcloud
nltk
vaderSentiment
python-dotenv
